+++ concordium-node/src/plugins/consensus.rs
@@ -11,7 +11,7 @@ use crate::{
         consensus::{
             self, ConsensusContainer, ConsensusRuntimeParameters, Regenesis, CALLBACK_QUEUE,
         },
-        ffi::{self, NotificationContext},
+        ffi::{self, ExecuteBlockCallback, NotificationContext},
         helpers::{
             ConsensusFfiResponse,
             PacketType::{self, *},
@@ -257,11 +257,9 @@ pub fn handle_consensus_inbound_msg(
 
     let source = request.source_peer();
     // relay external messages to Consensus
-    let consensus_result = send_msg_to_consensus(node, source, consensus, &request)?;
-
+    let (consensus_result, finalizer) = send_msg_to_consensus(node, source, consensus, &request)?;
     // adjust the peer state(s) based on the feedback from Consensus
     update_peer_states(node, &request, consensus_result);
-
     // rebroadcast incoming broadcasts if applicable
     if !drop_message
         && request.distribution_mode() == DistributionMode::Broadcast
@@ -275,6 +273,12 @@ pub fn handle_consensus_inbound_msg(
             (request.payload, request.variant),
         );
     }
+    // Execute any finalizer returned from the consensus layer.
+    if let Some(callback) = finalizer {
+        // Execute the block in the finalizer.
+        // There is nothing left to do afterwards.
+        let _ = consensus.execute_block(callback);
+    } // Else there is nothing to do, the block is processed.
 
     Ok(())
 }
@@ -284,38 +288,45 @@ fn send_msg_to_consensus(
     source_id: RemotePeerId,
     consensus: &ConsensusContainer,
     message: &ConsensusMessage,
-) -> anyhow::Result<ConsensusFfiResponse> {
+) -> anyhow::Result<(ConsensusFfiResponse, Option<ExecuteBlockCallback>)> {
     let payload = &message.payload[1..]; // non-empty, already checked
 
     let consensus_response = match message.variant {
-        Transaction => consensus.send_transaction(payload).1,
-        _ => {
+        Block => {
             let genesis_index = u32::deserial(&mut Cursor::new(&payload[..4]))?;
-            match message.variant {
-                Block => consensus.send_block(genesis_index, &payload[4..]),
-                FinalizationMessage => consensus.send_finalization(genesis_index, &payload[4..]),
-                FinalizationRecord => {
-                    consensus.send_finalization_record(genesis_index, &payload[4..])
-                }
-                CatchUpStatus => consensus.receive_catch_up_status(
+            consensus.receive_block(genesis_index, &payload[4..])
+        }
+        FinalizationMessage => {
+            let genesis_index = u32::deserial(&mut Cursor::new(&payload[..4]))?;
+            (consensus.send_finalization(genesis_index, &payload[4..]), Option::None)
+        }
+        FinalizationRecord => {
+            let genesis_index = u32::deserial(&mut Cursor::new(&payload[..4]))?;
+            (consensus.send_finalization_record(genesis_index, &payload[4..]), Option::None)
+        }
+        CatchUpStatus => {
+            let genesis_index = u32::deserial(&mut Cursor::new(&payload[..4]))?;
+            (
+                consensus.receive_catch_up_status(
                     genesis_index,
                     &payload[4..],
                     source_id,
                     node.config.catch_up_batch_limit,
                 ),
-                Transaction => unreachable!("Transaction message variant already handled."),
-            }
+                Option::None,
+            )
         }
+        Transaction => (consensus.send_transaction(payload).1, Option::None),
     };
 
-    if consensus_response.is_acceptable() {
+    if consensus_response.0.is_acceptable() {
         debug!("Processed a {} from {}", message.variant, source_id);
     } else {
         let num_bad_events = node.bad_events.inc_invalid_messages(source_id);
         // we do log some invalid messages to both ease debugging and see problems in
         // normal circumstances
         if num_bad_events < 10 {
-            warn!("Couldn't process a {} due to error code {:?}", message, consensus_response);
+            warn!("Couldn't process a {} due to error code {:?}", message, consensus_response.0);
         }
     }