+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## Unreleased changes
+- Relay blocks earlier. In particular this means that blocks are now processed in 
+  two steps, `block receive` and `block execute`. The former performs verification of block meta data
+  while the latter adds the block to the tree.
+  Blocks are now enqueued in the outgoing message queue in between the the two steps.
 - Removed the configuration option 'no_rebroadcast_consensus_validation'.
 
 ## 5.1.0